# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/itemku/transactions/status": {
      "post": {
        "description": "Re-reads the supplier status for one reference id. NOTE: this route carries @UserRoles but no @UseGuards(AuthGuard), so no token is checked today and the role restriction is not enforced.",
        "operationId": "TransactionController_getPaymentTransactionStatus",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemkuRefIdDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Status of a payment transaction",
        "tags": [
          "Itemku Transaction"
        ]
      }
    },
    "/api/v1/itemku/transactions/update-all-status": {
      "post": {
        "description": "Walks all non-final transactions and refreshes them from the supplier. No body. NOTE: this route carries @UserRoles but no @UseGuards(AuthGuard), so no token is checked today and the role restriction is not enforced.",
        "operationId": "TransactionController_updateAllStatus",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Re-sync every pending transaction",
        "tags": [
          "Itemku Transaction"
        ]
      }
    },
    "/api/v1/itemku/transactions/{id}": {
      "get": {
        "description": "Returns the state of the BullMQ job that carries this transaction.",
        "operationId": "TransactionController_getJobTransactionStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Queue job id",
            "schema": {
              "example": "1421",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Queue job status",
        "tags": [
          "Itemku Transaction"
        ]
      }
    },
    "/api/v1/itemku/transactions/retry": {
      "post": {
        "description": "SuperAdmin only. Re-queues the given Itemku order ids.",
        "operationId": "TransactionController_createPaymentTransactionRequest",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetryItemkuOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Retry failed Itemku orders",
        "tags": [
          "Itemku Transaction"
        ]
      }
    },
    "/api/v1/itemku/transactions/compare": {
      "post": {
        "description": "Resolves the current selling price of a SKU for a customer and checks it against the supplied balance. NOTE: this route carries @UserRoles but no @UseGuards(AuthGuard), so no token is checked today and the role restriction is not enforced.",
        "operationId": "TransactionController_compareCurrentPrice",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComparePriceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Compare current price against a balance",
        "tags": [
          "Itemku Transaction"
        ]
      }
    },
    "/api/v1/itemku/transactions/get-one": {
      "post": {
        "description": "Looks up one transaction history record by reference id.",
        "operationId": "TransactionController_getTxHistory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemkuRefIdDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Single transaction history entry",
        "tags": [
          "Itemku Transaction"
        ]
      }
    },
    "/api/v1/itemku/transactions/get": {
      "post": {
        "description": "Paginated history. Note that `status` and `source` are not combined -- when `source` is present it replaces the `status` filter.",
        "operationId": "TransactionController_getTxHistories",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetTxHistoriesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Transaction history list",
        "tags": [
          "Itemku Transaction"
        ]
      }
    },
    "/api/v1/itemku/transactions/get-itemku-history": {
      "post": {
        "description": "Paginated Itemku order list. Text filters match partially and case insensitively; the date range applies only when both `dateStart` and `dateEnd` are given.",
        "operationId": "TransactionController_getItemkuTxHistories",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetTransaction"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Itemku order history",
        "tags": [
          "Itemku Transaction"
        ]
      }
    },
    "/api/v1/itemku/transactions/mock-order-itemku": {
      "post": {
        "description": "SuperAdmin only. Feeds a synthetic Itemku order through the same path as `/update-order-itemku`. Intended for testing.",
        "operationId": "TransactionController_mockItemkuOrder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemkuOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Simulate an Itemku order webhook",
        "tags": [
          "Itemku Transaction"
        ]
      }
    },
    "/api/v1/itemku/transactions/update-order-itemku": {
      "post": {
        "description": "Stores the order and moves its transaction forward when the status changed.",
        "operationId": "TransactionController_updateItemkuOrder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemkuOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Upsert an Itemku order",
        "tags": [
          "Itemku Transaction"
        ]
      }
    },
    "/api/v1/itemku/transactions/bulk-update-tx": {
      "post": {
        "description": "Refreshes every tracked Itemku order in one pass. No body.",
        "operationId": "TransactionController_bulkUpdatetx",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Bulk re-sync Itemku orders",
        "tags": [
          "Itemku Transaction"
        ]
      }
    },
    "/api/v1/itemku/transactions/manual-update-tx": {
      "post": {
        "description": "Refreshes the given reference ids against the supplier.",
        "operationId": "TransactionController_manualUpdatetx",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualUpdateTxDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Manually re-sync transactions",
        "tags": [
          "Itemku Transaction"
        ]
      }
    },
    "/api/v1/itemku/transactions/get-digiflazz": {
      "post": {
        "description": "SuperAdmin only. Returns the raw supplier price list. No body.",
        "operationId": "TransactionController_getDigiflazzProductPrice",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Fetch the Digiflazz price list",
        "tags": [
          "Itemku Transaction"
        ]
      }
    },
    "/api/v1/itemku/transactions/get-itemku": {
      "post": {
        "description": "SuperAdmin only. Returns the raw Itemku price list. No body.",
        "operationId": "TransactionController_getItemkuProductPrice",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Fetch the Itemku price list",
        "tags": [
          "Itemku Transaction"
        ]
      }
    },
    "/api/v1/itemku/transactions/update-tx-history": {
      "post": {
        "description": "Rewrites the SKU / destination of an existing transaction, matched on `ref_id`.",
        "operationId": "TransactionController_updateTransactionHistoryDetail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTransactionRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update a transaction history entry",
        "tags": [
          "Itemku Transaction"
        ]
      }
    },
    "/api/v1/itemku/transactions/itemku/product/upload": {
      "post": {
        "description": "Multipart upload of a single `.xls`/`.xlsx` file (10 MB max). The sheet is parsed and the rows are imported. Rejects any other extension with a 400.",
        "operationId": "TransactionController_uploadFile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Excel file (.xls or .xlsx)"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Import Itemku products from Excel",
        "tags": [
          "Itemku Transaction"
        ]
      }
    },
...
  },
  "info": {
    "title": "TRX API Service",
    "description": "Transaction / top-up API service.\n\n## Authentication\n\nCall `POST /api/v1/auth/login` to obtain a JWT, then send it in the\n`authorization` header **without** the `Bearer ` prefix -- the guard verifies\nthe header value verbatim.\n\nMachine-to-machine endpoints (those under `/api/v1/tx/ext/...`) accept the\n`x-username` + `x-sign` header pair instead of a JWT. `x-sign` is the\nHMAC-SHA256 hex digest of the exact JSON request body signed with the\ncaller's API key (see `POST /api/v1/auth/api-key`).\n\nA few endpoints (`/api/v1/tx/request-web`, `/api/v1/external-topup/create-topup`,\n`/api/v1/external-topup/update-topup`) require BOTH a JWT and an `x-sign`\nheader over the body.\n\n## Response envelope\n\nMost endpoints answer with `{ message, data, errorCode }`. An `errorCode` of\n`\"0000\"` means success; any other code maps to an entry in\n`src/lib/errorCodes.ts` (e.g. `1000` wrong credential, `1002` insufficient\nbalance, `4012` signature not verified).",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "type": "apiKey",
        "in": "header",
        "name": "authorization",
        "description": "Raw JWT returned by POST /api/v1/auth/login. Do NOT prefix it with \"Bearer \"."
      },
      "x-username": {
        "type": "apiKey",
        "in": "header",
        "name": "x-username",
        "description": "Username of the API consumer. Used together with x-sign."
      },
      "x-sign": {
        "type": "apiKey",
        "in": "header",
        "name": "x-sign",
        "description": "HMAC-SHA256 hex signature of the JSON request body."
      }
    },
    "schemas": {
      "ItemkuRefIdDto": {
        "type": "object",
        "properties": {
          "ref_id": {
            "type": "string",
            "example": "TRX-20250901-0001",
            "description": "Transaction reference id"
          }
        },
        "required": [
          "ref_id"
        ]
      },
      "ApiResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "success",
            "description": "Human readable result message"
          },
          "data": {
            "type": "object",
            "nullable": true,
            "description": "Endpoint specific payload. `null` when the call failed."
          },
          "errorCode": {
            "type": "string",
            "example": "0000",
            "description": "`0000` on success, otherwise a code from src/lib/errorCodes.ts"
          }
        },
        "required": [
          "message",
          "data"
        ]
      },
      "RetryItemkuOrderDto": {
        "type": "object",
        "properties": {
          "order_id": {
            "example": [
              90210,
              90211
            ],
            "description": "Itemku order ids to re-run",
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "order_id"
        ]
      },
      "ComparePriceDto": {
        "type": "object",
        "properties": {
          "buyer_sku_code": {
            "type": "string",
            "example": "ml86",
            "description": "Supplier SKU code to price"
          },
          "userBalance": {
            "type": "number",
            "example": 500000,
            "description": "Balance to check the price against"
          },
          "username": {
            "type": "string",
            "example": "johndoe",
            "description": "Customer the price applies to"
          }
        },
        "required": [
          "buyer_sku_code",
          "userBalance",
          "username"
        ]
      },
      "GetTxHistoriesDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1,
            "description": "Page number, 1 based"
          },
          "take": {
            "type": "number",
            "example": 10,
            "description": "Rows per page"
          },
          "status": {
            "type": "string",
            "example": "SUCCESS",
            "description": "Filter by transaction status"
          },
          "source": {
            "type": "string",
            "example": "itemku",
            "description": "Filter by origin. Overrides `status` -- only one filter is applied."
          }
        }
      },
      "GetTransaction": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1,
            "description": "Page number, 1 based"
          },
          "size": {
            "type": "number",
            "example": 10,
            "description": "Rows per page"
          },
          "sort": {
            "type": "string",
            "example": "desc",
            "description": "Sort direction on the update timestamp"
          },
          "dateStart": {
            "type": "string",
            "example": "2025-09-01",
            "description": "Inclusive lower bound. Applied only together with dateEnd."
          },
          "dateEnd": {
            "type": "string",
            "example": "2025-09-30",
            "description": "Inclusive upper bound. Applied only together with dateStart."
          },
          "status": {
            "type": "string",
            "example": "SUCCESS",
            "description": "Order status"
          },
          "deliveryStatus": {
            "type": "string",
            "example": "delivered",
            "description": "Delivery status, matched case insensitively"
          },
          "gameName": {
            "type": "string",
            "example": "Mobile Legends",
            "description": "Partial, case insensitive match"
          },
          "productName": {
            "type": "string",
            "example": "86 Diamonds",
            "description": "Partial, case insensitive match"
          },
          "orderId": {
            "type": "string",
            "example": "90210",
            "description": "Exact numeric order id"
          },
          "orderNumber": {
            "type": "string",
            "example": "INV-20250901-001",
            "description": "Partial, case insensitive match"
          },
          "userInfo": {
            "type": "string",
            "example": "12345678(1234)",
            "description": "Buyer / destination account info"
          }
        }
      },
...
      "ManualUpdateTxDto": {
        "type": "object",
        "properties": {
          "refIds": {
            "example": [
              "TRX-20250901-0001",
              "TRX-20250901-0002"
            ],
            "description": "Reference ids to re-sync against the supplier",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "refIds"
        ]
      },
      "UpdateTransactionRequestDto": {
        "type": "object",
        "properties": {
          "buyer_sku_code": {
            "type": "string",
            "example": "ml86",
            "description": "Supplier SKU code of the transaction"
          },
          "customer_no": {
            "type": "string",
            "example": "12345678(1234)",
            "description": "Destination account"
          },
          "source": {
            "type": "string",
            "example": "itemku",
            "description": "Origin of the transaction"
          },
          "ref_id": {
            "type": "string",
            "example": "TRX-20250901-0001",
            "description": "Reference id of the transaction to update"
          }
        },
        "required": [
          "buyer_sku_code",
          "customer_no",
          "ref_id"
        ]
      },
...
    }
  }
}